cleaning

+++ src/main/java/com/marginallyclever/donatello/nodes/images/blend/BlendScreen.java
@@ -32,8 +32,8 @@ public class BlendScreen extends Node {
         BufferedImage A = a.getValue();
         BufferedImage B = b.getValue();
 
-        int w = (int)Math.min(A.getWidth(),B.getWidth());
-        int h = (int)Math.min(A.getHeight(),B.getHeight());
+        int w = Math.min(A.getWidth(),B.getWidth());
+        int h = Math.min(A.getHeight(),B.getHeight());
         BufferedImage C = new BufferedImage(w,h,A.getType());
 
         for(int y=0;y<h;++y) {
